feat: export finder helpers
