+++ src/utils/finders.ts
@@ -1,6 +1,17 @@
 import type { SnapshotNode } from './snapshot.ts';
+import { AppError } from './errors.ts';
 
 export type FindLocator = 'any' | 'text' | 'label' | 'value' | 'role' | 'id';
+
+export type FindAction =
+  | { kind: 'click' }
+  | { kind: 'focus' }
+  | { kind: 'fill'; value: string }
+  | { kind: 'type'; value: string }
+  | { kind: 'get_text' }
+  | { kind: 'get_attrs' }
+  | { kind: 'exists' }
+  | { kind: 'wait'; timeoutMs?: number };
 
 type FindMatchOptions = {
   requireRect?: boolean;
@@ -89,10 +100,60 @@ export function normalizeText(value: string): string {
   return value.trim().toLowerCase().replace(/\s+/g, ' ');
 }
 
-function normalizeRole(value: string): string {
+export function normalizeRole(value: string): string {
   let normalized = value.trim();
   if (!normalized) return '';
   const lastSegment = normalized.split('.').pop() ?? normalized;
   normalized = lastSegment.replace(/XCUIElementType/gi, '').toLowerCase();
   return normalized;
+}
+
+export function parseFindArgs(args: string[]): {
+  locator: FindLocator;
+  query: string;
+  action: FindAction['kind'];
+  value?: string;
+  timeoutMs?: number;
+} {
+  const locatorTokens: FindLocator[] = ['text', 'label', 'value', 'role', 'id'];
+  let locator: FindLocator = 'any';
+  let queryIndex = 0;
+  if (locatorTokens.includes(args[0] as FindLocator)) {
+    locator = args[0] as FindLocator;
+    queryIndex = 1;
+  }
+  const query = args[queryIndex] ?? '';
+  const actionTokens = args.slice(queryIndex + 1);
+  if (actionTokens.length === 0) {
+    return { locator, query, action: 'click' };
+  }
+  const action = actionTokens[0].toLowerCase();
+  if (action === 'get') {
+    const sub = actionTokens[1]?.toLowerCase();
+    if (sub === 'text') return { locator, query, action: 'get_text' };
+    if (sub === 'attrs') return { locator, query, action: 'get_attrs' };
+    throw new AppError('INVALID_ARGS', 'find get only supports text or attrs');
+  }
+  if (action === 'wait') {
+    const timeoutMs = parseTimeout(actionTokens[1]);
+    return { locator, query, action: 'wait', timeoutMs: timeoutMs ?? undefined };
+  }
+  if (action === 'exists') return { locator, query, action: 'exists' };
+  if (action === 'click') return { locator, query, action: 'click' };
+  if (action === 'focus') return { locator, query, action: 'focus' };
+  if (action === 'fill') {
+    const value = actionTokens.slice(1).join(' ');
+    return { locator, query, action: 'fill', value };
+  }
+  if (action === 'type') {
+    const value = actionTokens.slice(1).join(' ');
+    return { locator, query, action: 'type', value };
+  }
+  throw new AppError('INVALID_ARGS', `Unsupported find action: ${actionTokens[0]}`);
+}
+
+function parseTimeout(value: string | undefined): number | null {
+  if (!value) return null;
+  const parsed = Number(value);
+  return Number.isFinite(parsed) ? parsed : null;
 }
